+++ muk_rest/controllers/system.py
@@ -65,7 +65,7 @@ class SystemController(http.Controller):
     # System
     #----------------------------------------------------------
 
-    @http.route('/api/modules', auth="none", type='http', methods=['GET'], csrf=False)
+    @http.route('/api/modules', auth="none", type='http', methods=['GET'])
     @tools.common.parse_exception
     @tools.common.ensure_database
     def modules(self):
@@ -76,14 +76,14 @@ class SystemController(http.Controller):
         content = json.dumps(topological_sort(result), sort_keys=True, indent=4, cls=ResponseEncoder)
         return Response(content, content_type='application/json;charset=utf-8', status=200)
     
-    @http.route('/api/languages', auth="none", type='http', methods=['GET'], csrf=False)
+    @http.route('/api/languages', auth="none", type='http', methods=['GET'])
     @tools.common.parse_exception
     def languages(self):
         result = [(lang[0], lang[1].split("/")[0].strip()) for lang in service.db.exp_list_lang()]
         content = json.dumps(result, sort_keys=True, indent=4, cls=ResponseEncoder)
         return Response(content, content_type='application/json;charset=utf-8', status=200)
     
-    @http.route('/api/countries', auth="none", type='http', methods=['GET'], csrf=False)
+    @http.route('/api/countries', auth="none", type='http', methods=['GET'])
     @tools.common.parse_exception
     def countries(self):
         result = service.db.exp_list_countries()