+++ muk_rest/models/oauth1.py
@@ -100,7 +100,6 @@ class OAuth1(models.Model):
     # Create / Update / Delete
     #----------------------------------------------------------
 
-    @api.multi
     def unlink(self):
         self.mapped('oauth').unlink()
         return super(OAuth1, self).unlink()