+++ muk_rest/tests/test_custom.py
@@ -58,10 +58,9 @@ from odoo.addons.muk_rest.tests.common import active_authentication, RestfulCase
 _path = os.path.dirname(os.path.dirname(__file__))
 _logger = logging.getLogger(__name__)
 
-CUSTOM_URL = '/api/custom'
-CUSTOM_DOMAIN_URL = '/domain'
-CUSTOM_ACTION_URL = '/action'
-CUSTOM_CODE_URL = '/code'
+CUSTOM_DOMAIN_URL = 'test_domain'
+CUSTOM_ACTION_URL = 'test_action'
+CUSTOM_CODE_URL = 'test_code'
 
 class CustomTestCase(RestfulCase):
     
@@ -73,17 +72,20 @@ class CustomTestCase(RestfulCase):
             'model': self.ref('base.model_res_partner'),
             'method': 'GET',
             'state': 'domain'})
-        action = self.env['ir.actions.server'].create({
+        action_vals = {
             'name': 'Action Action',
             'model_id': self.ref('base.model_res_partner'),
             'code': "log('testing')",
-            'activity_user_type': 'specific',
-            'state': 'code'})
+            'state': 'code'
+        }
+        action_model = self.env['ir.actions.server']
+        if 'activity_user_type' in action_model._fields :
+            action_vals.update({'activity_user_type': 'specific'})
         self.action_endpoint = self.env['muk_rest.endpoint'].create({
             'name': 'Action Test',
             'endpoint': CUSTOM_ACTION_URL,
             'model': self.ref('base.model_res_partner'),
-            'action': action.id,
+            'action': action_model.create(action_vals).id,
             'method': 'POST',
             'state': 'action'})
         self.code_endpoint = self.env['muk_rest.endpoint'].create({
@@ -97,7 +99,7 @@ class CustomTestCase(RestfulCase):
     @unittest.skipIf(not active_authentication, "Skipped because no authentication is available!")
     def test_domain(self):
         client = self.authenticate()
-        response = client.get(self.url_prepare(CUSTOM_URL + CUSTOM_DOMAIN_URL))
+        response = client.get(self.url_prepare(self.domain_endpoint.route))
         self.assertTrue(response)
         self.assertTrue(response.json())
         
@@ -105,7 +107,7 @@ class CustomTestCase(RestfulCase):
     def test_domain_simple(self):
         client = self.authenticate()
         self.domain_endpoint.write({'domain': '[["id","=",1]]'})
-        response = client.get(self.url_prepare(CUSTOM_URL + CUSTOM_DOMAIN_URL))
+        response = client.get(self.url_prepare(self.domain_endpoint.route))
         self.assertTrue(response)
         self.assertTrue(response.json())
         
@@ -113,7 +115,7 @@ class CustomTestCase(RestfulCase):
     def test_domain_field(self):
         client = self.authenticate()
         self.domain_endpoint.write({'domain_fields': [(6, 0, [self.ref('base.field_res_partner__name')])]})
-        response = client.get(self.url_prepare(CUSTOM_URL + CUSTOM_DOMAIN_URL))
+        response = client.get(self.url_prepare(self.domain_endpoint.route))
         self.assertTrue(response)
         self.assertTrue(response.json())
         
@@ -121,27 +123,27 @@ class CustomTestCase(RestfulCase):
     def test_domain_context(self):
         client = self.authenticate()
         self.domain_endpoint.write({'domain': '[["id","=",active_id]]'})
-        response = client.get(self.url_prepare(CUSTOM_URL + CUSTOM_DOMAIN_URL), data={'id': 1})
+        response = client.get(self.url_prepare(self.domain_endpoint.route), data={'id': 1})
         self.assertTrue(response)
         self.assertTrue(response.json())
         
     @unittest.skipIf(not active_authentication, "Skipped because no authentication is available!")
     def test_domain_demo(self):
         client = self.authenticate("demo", "demo")
-        response = client.get(self.url_prepare(CUSTOM_URL + CUSTOM_DOMAIN_URL))
+        response = client.get(self.url_prepare(self.domain_endpoint.route))
         self.assertTrue(response)
         self.assertTrue(response.json())
         
     @unittest.skipIf(not active_authentication, "Skipped because no authentication is available!")
     def test_action(self):
         client = self.authenticate()
-        response = client.get(self.url_prepare(CUSTOM_URL + CUSTOM_ACTION_URL))
+        response = client.get(self.url_prepare(self.action_endpoint.route))
         self.assertTrue(response)
         self.assertTrue(response.json())
         
     @unittest.skipIf(not active_authentication, "Skipped because no authentication is available!")
     def test_code(self):
         client = self.authenticate()
-        response = client.get(self.url_prepare(CUSTOM_URL + CUSTOM_CODE_URL))
+        response = client.get(self.url_prepare(self.code_endpoint.route))
         self.assertTrue(response)
         self.assertTrue(response.json())